Crate finality_relay

This crate has single entrypoint to run synchronization loop that is built around finality proofs, as opposed to headers synchronization loop, which is built around headers. The headers are still submitted to the target node, but are treated as auxiliary data as we are not trying to submit all source headers to the target node.

Structs

FinalitySyncParams

Finality proof synchronization loop parameters.

Traits

FinalitySyncPipeline

Finality proofs synchronization pipeline.

SourceClient

Source client used in finality synchronization loop.

SourceHeader

Header that we’re receiving from source node.

TargetClient

Target client used in finality synchronization loop.

Functions

metrics_prefix

Return prefix that will be used by default to expose Prometheus metrics of the finality proofs sync loop.

run

Run finality proofs synchronization loop.